Silenced warnings and updated the MSVC project file
git-svn-id: https://svn.scummvm.org:4444/svn/comet@57 e395a8f9-5cf0-422e-971d-9a083d8f24e8

+++ engines/comet/dialog.cpp
@@ -135,7 +135,7 @@ void Dialog::drawTextBubbles() {
 
 		color2 = color1;
 
-		if (i == _dialogSelectedItemIndex) {
+		if (i == (uint)_dialogSelectedItemIndex) {
 			if (_vm->_sceneObjects[0].textColor == 25) {
 				color2 = _dialogTextColor;
 				_dialogTextColor += _dialogTextColorInc;
